test: fix broken vitest unit tests
