+++ solution/ui/regulations/eregs-vite/src/components/subjects/PolicyResults.test.js
@@ -363,132 +363,3 @@ describe("showResultSnippet", () => {
         expect(PolicyResults.showResultSnippet(cs_external_2)).toBe(false);
     });
 });
-
-describe("getCollapseName", () => {
-    const doc1 = {
-        node_id: "doc1-node_id",
-        id: "doc1-id",
-    };
-
-    it("returns the id if id exists", async () => {
-        expect(PolicyResults.getCollapseName(doc1)).toBe(
-            "related citations collapsible doc1-id"
-        );
-    });
-
-    const doc2 = {
-        node_id: "doc2-node_id",
-    };
-
-    it("returns the node_id if no id", async () => {
-        expect(PolicyResults.getCollapseName(doc2)).toBe(
-            "related citations collapsible doc2-node_id"
-        );
-    });
-});
-
-describe("hasRegulationCitations", () => {
-    const partsLastUpdated = {
-        438: "2024-07-09",
-    };
-
-    const doc1 = {
-        cfr_citations: [{ title: "42", part: "438" }],
-    };
-
-    it("returns true if the document has cfr_citations with an updated part", async () => {
-        expect(
-            PolicyResults.hasRegulationCitations({
-                doc: doc1,
-                partsLastUpdated,
-            })
-        ).toBe(true);
-    });
-
-    const doc2 = {
-        cfr_citations: [{ title: "42", part: "439" }],
-    };
-
-    it("returns false if the document has cfr_citations with an unknown part", async () => {
-        expect(
-            PolicyResults.hasRegulationCitations({
-                doc: doc2,
-                partsLastUpdated,
-            })
-        ).toBe(false);
-    });
-
-    const doc3 = {
-        cfr_citations: [],
-    };
-
-    it("returns false if the document has no cfr_citations", async () => {
-        expect(
-            PolicyResults.hasRegulationCitations({
-                doc: doc3,
-                partsLastUpdated,
-            })
-        ).toBe(false);
-    });
-});
-
-describe("hasStatuteCitations", () => {
-    it("returns false if the document has no act_citations or no usc_citations", async () => {
-        expect(
-            PolicyResults.hasStatuteCitations({
-                doc: {
-                    id: 1,
-                    unrelated_field: ["x"],
-                },
-            })
-        ).toBe(false);
-
-        expect(
-            PolicyResults.hasStatuteCitations({
-                doc: {
-                    id: 2,
-                    act_citations: [],
-                    usc_citations: [],
-                },
-            })
-        ).toBe(false);
-
-        expect(
-            PolicyResults.hasStatuteCitations({
-                doc: {
-                    id: 3,
-                    act_citations: [1, 2],
-                },
-            })
-        ).toBe(false);
-
-        expect(
-            PolicyResults.hasStatuteCitations({
-                doc: {
-                    id: 4,
-                    act_citations: [],
-                },
-            })
-        ).toBe(false);
-
-        expect(
-            PolicyResults.hasStatuteCitations({
-                doc: {
-                    id: 5,
-                    usc_citations: [1, 2],
-                    act_citations: [],
-                },
-            })
-        ).toBe(true);
-
-        expect(
-            PolicyResults.hasStatuteCitations({
-                doc: {
-                    id: 6,
-                    act_citations: [1, 2],
-                    usc_citations: ["one", "two"],
-                },
-            })
-        ).toBe(true);
-    });
-});
